requirements.txt
Adafruit-Blinka==8.56.0
Adafruit-BNO055==1.0.2
adafruit-circuitpython-bme280==2.6.27
adafruit-circuitpython-bme680==3.7.12
adafruit-circuitpython-bno055==5.4.18
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-gps==3.11.1
adafruit-circuitpython-ina219==3.4.26
adafruit-circuitpython-register==1.10.2
adafruit-circuitpython-requests==4.1.10
adafruit-circuitpython-typing==1.11.2
Adafruit-GPIO==1.0.3
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11
attrs==25.1.0
av==14.2.0
binho-host-adapter==0.1.6
board==1.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
libarchive-c==5.1
MarkupSafe==3.0.2
mock==5.2.0
numpy==2.2.3
pi-ina219==1.4.1
picamera2==0.3.25
pidng==4.0.9
piexif==1.1.3
pigpio==1.78
pillow==11.1.0
ply==3.11
pyftdi==0.56.0
pynmea2==1.19.0
pyserial==3.5
python-prctl==1.8.1
pyusb==1.3.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
simplejpeg==1.8.2
smbus2==0.5.0
spidev==3.6
sysv_ipc==1.1.0
tqdm==4.67.1
typing_extensions==4.13.1
v4l2-python3==0.3.5
